TABLE UPDATES AND R CONNECT SCRIPT
The table creation script now uses BigInt for the finance table. Could have been more subtle but for a db of this size didn't see the need to trawl through the data to check.

There's now a connect script to source in R so I can quickly query data into dataframes.